Show attribute name instead of foreign key (number) itself
This commit lets us to see readable name instead of internal primary key
(id/number). For example, "color: 1" in the previous commit is changed
to "color: brown".

This commit drops support of original `poo` table in actual database.
It's because the users don't have to take care of original id/name
combination used in database.