Add test code for index page
Since the app config is common between `cargo run` and `cargo test`,
this commit adds function `app_config` to main module.

Tests using database must NOT be run in parallel. If so, data stored
in DB could be unstable. To work with this issue, this commit adds
mutex lock to `run_db_test` macro. Note that index page test doesn't
use this macro because it doesn`t test db contents.

This commit resolves #2.